download Devise branch from github
this is required to sidestep the fact that Devise has compatible Omniauth
versions hardcoded as '1.x.x' and Omniauth has recently gone '2.0.x'.

the Devise maintainer is planning a release that will include the fix for
this, in the meantime advised for people to keep using this branch.
See https://github.com/heartcombo/devise/pull/5327 for more details.