build(docker): fix image build failing
Installing bundler deps with the app user was failing with a permission
error, I think it was trying to modify the root-owned gems related to
nokogiri installed prior to the `bundle install` step. Installing
nokogiri there was not required, as it's a dependency of other gems in
our Gemfile.

I've also restructured the steps to only do `bundle install` once, and
the app files are copied after the bundle install to avoid invalidating
the build cache when files change.