fix: allow saving Spaces that create new Libraries
Saving Spaces was also failing, due to Rails 5 belongs_to non-optional
validation. The Library given by new_library_name was being created
in Space.before_save(), but that is called after validation, so the
Space would be invalid due to having no value for the library relation.
To fix this, an un-saved Library is now assigned to the library
relation when the Space is created, and Active Record Autosave handles
saving both when the Space is created.