feat: enable a mock authentication provider in development
When the app is run in development mode, the omniauth developer provider
is enabled. This is a mock provider that just allows you to enter the
user details you want in a form, and creates/logs you in with those
details. This avoids the need to set up SAML auth for a dev environment.