fix: work around paperclip crashing
The paperclip gem is deprecated and no longer updated. It uses an API
that was removed in Ruby 3 (URI.escape), causing it to crash when
calling url() on a paperclip attachment object. As a work around, the
'escape' option to url() can be used to disable paperclip's faulty URL
escaping code.