build(devcontainer): don't fail to start when bin/setup fails
The VSCode treats the onCreateCommand failing as a sign that the
devcontainer environment is unusable, and puts you into a recovery
container. In our case it's not such a big problem, so onCreateCommand
now handles errors from bin/setup by reporting them in log output,
allowing the devcontainer to open.