chore(devcontainer): support local and remote
Support running the devcontainer with a local bind mount as well as
when running from a container volume.

The way this works is a little bit of a hack, but not too much. 😊
When running from a container volume, ${localWorkspaceFolder} is not
set. And the workspaceFolder devcontainer.json setting is ignored.
We put the value of ${localWorkspaceFolder} in a .env file as
LOCAL_WORKSPACE_FOLDER for docker-compose to read. In the compose file
we use the value of LOCAL_WORKSPACE_FOLDER to only set up a bind mount
if it's set.

So when using a container volume, the /workspaces/$name dir is
automatically selected as the workspaceFolder by VSCode. In a local bind
mount /workspace is used, and it's bound to the local dir.

A final quirk is that when ${localWorkspaceFolder} is not set, it
evalutes to literally '${localWorkspaceFolder}'. Happilly docker-compose
tries to evalute that as an environment variable, which turns it back
into the empty string.