fix: load Google Maps with empty API key
The Google Maps JS API was failing to load on localhost due to the API
key not permitting that origin. With an empty API key it loads and works
as expected on localhost. The SPACEFINDER_GOOGLE_MAPS_API_KEY envar can
be used to set the API key for a production deployment.